Hay un problema de incializacion en cargar_cms de couint_min_sketch. Raro.

+++ procesar_tweets.h
@@ -1,7 +1,13 @@
 #ifndef PROCESAR_TWEETS
 #define PROCESAR_TWEETS
-#include "stdlib.h"
-#include "stdio.h"
+#include <stdlib.h>
+#include <stdio.h>
+#include <stdbool.h>
+#include <string.h>
+#include "tools/count_min_sketch.h"
+#include "tools/strutil.h"
+#include "tools/hash.h"
+
 
 //Esta estructura contiene el hashtag y sus apariciones
 typedef struct hashtag ht_t;
@@ -15,6 +21,6 @@ int cmp_hashtag(const void* uno, const void* dos);
 
 
 //Devuelve cada k lineas ingresadas por entrada estandar los n hashtags mas frecuentes
-void procesar_tweets(int n, int k);
+void procesar_tweets(size_t n, int k);
 
 #endif  // PROCESAR_TWEETS
